refactor(cli): move schedule sub-app to app/cli/schedule.py
Moves the four schedule_* commands plus _load_schedule_payload and
_resolve_schedule out of app/main.py. main.py temporarily re-imports
schedule_list_cmd because scheduler_status_cmd (still in main.py) calls
it directly; Task 10 will move scheduler_status_cmd and absorb this
import. Drops orphaned imports for schedule-only deps.
